Generate, build, and package API client libraries
This change brings in a copy of the google/apis-client-generator code
used to generate per-API libraries, and wires it up with MSBuild.

The process is orchestrated by GeneratedLibraries.proj:
    msbuild /m GeneratedLibraries.proj

First, we download the discovery documents. This is isolated as its own
step so we have copies of the actual discovery documents used to produce
the generated libraries. This leaves the rest of the build hermetic and
will help us reproduce and fix codegen problems in the future.

Next, the client generator is run on each discovery document. The generated
source ends up at Src/Generated/<API PACKAGE NAME>.

Each client library project is built. The GeneratedLibraries.proj includes
a list of the projects that don't currently build so they can be ignored
for now.

Finally, we run NuGet to package each of the libraries. The packages are
put into NuPkgs/Generated.