Provide synchronous method to obtain default application credentials
This addresses issue #652.

The "remarks" comment is deliberately off-putting, in that if you're in a situation where you *can*
sensibly use the async call, you should do so - but as the rest of the client library allows synchronous
calls, it would be odd not to do so for initialization.

Alternatives to consider:
- Put synchronous code in DefaultCredentialProvider (no obvious benefit, as we'd still need to await an HttpClient eventually)
- Just document the "safety" of using the Result property in the async call