Add discovery documents and generated code to the repository.
As generated by:
  msbuild /nr:false /m GeneratedLibraries.proj

By keeping a consistent, point-in-time view of the API surface and the
libraries we generated from it, we'll make it much easier to isolate
and fix code generation bugs in the future. We can also catch possible
breaking changes in the generated code. Finally, it'll be nice to have
the code browsable on GitHub.

This was the state of the world at commit time. This isn't the code
for the versions of the packages currently on NuGet.

In the future, I expect we'll have automation to update these
regularly. Such automation should mark its commits somehow to allow
filtering them out when viewing history.