Capitalize packagePath from discovery document
The Generate function in generate_library.py sets the package_path
option to None by default. When the C# code generator looks for
package_path in order to capitalize its components, it finds the
None value and stops, instead of reading the packagePath value from
the discovery document.

Now, if package_path is None, we read the packagePath from the
discovery document.

When this code is used by Google's code generator server, the real
value of package_path is used by default, instead of None, and the
C# code actually capitalizes the library components. For example,
Google.Apis.Admin.DataTransfer.datatransfer.v1.

https://www.nuget.org/packages/Google.Apis.Admin.DataTransfer.datatransfer_v1

(The last component of *that* name, in turn, is lowercase in both
cases. This is likely a bug, but one that will remain for backward
compatibility.)