Update discovery documents and generated libraries
With this change, we should (finally) be able to publish the packages
from NuPkgs\Generated and retire the Google.Apis.NuGet.Publisher tool.

Version is incremented to 1.10.1.X to accommodate the new way of
generating the revision number.

Updated the .nuspec to point customers to the generated source on GitHub.

I've compared the assemblies produced in NuPkgs\Generated at this commit
with the ones generated by Google.Apis.NuGet.Publisher. The only difference
is in Google.Apis.TagManager.v1, and that was due to outdated code served
from http://google-api-client-libraries.appspot.com.