NOTE: This change depends on changes to https://github.com/google/google-api-cpp-client. The two repos may be out of sync for a few hours. They may be out of sync with the libraries published on https://developers.google.com/ for a few days.
Move the building of the full URL for a method from the caller site of
creation of the method instance to the callee site. This effectively
removes a call of the form StrCat(service->url, constant) from each
method in the API and replaces it with a single StrCat() for the entire
library. For x86_64, this is approx 100bytes per method. For a large API,
this can be significant. E.g. in Google Drive, the .o file reduces by 22K.