ReleaseTravis should use ReleaseSigned for projects it builds
When the solution builds the ReleaseTravis target, the new Net45
projects build for ReleaseSigned, but all existing projects were
configured to build in Release. The new Net45 projects failed to
compile because they were taking a dependency on non-strongnamed
assemblies.

Now that the signing key is available in GitHub, Travis should
build everything in ReleaseSigned.