Build the generated library PCLs for two different profiles
Since NuGet doesn't currently filter libraries based on their supported
target frameworks, these changes are necessary to ensure that
`Install-Package Google.Apis.Storage.v1`
still works in projects targeting .NET 4.0 now that new versions of the
support libraries no longer target .NET 4.0.

Profile259 supports newer frameworks and brings in the newest version
of the support libraries.

Profile328 is the "legacy" framework set and brings in the last version
of the support libraries that supported Net40: v1.10, referred to here
as oldCompatibleVersion for easier searchability.

Updated GeneratedLibraries.proj so it will build both projects.

Updated the .nuspec to include both built assemblies in the appropriate
portable-xyz directories.

Verified using test projects that the targetFramework values in
packages.config and the HintPaths in the .csprojs are correct
for the new project targets.