Avoid unnecessary copies by setting references to "not private"
Private references are copied to the bin\ directory as part of build. That's
useful when the directory will be copied for deployment; less so when we're
only interested in adding the main assembly to a NuGet package.

Setting `<Private>False</Private>` avoids unnecessarily copying hundreds of
DLLs during the build.