Resume upload after program restart - Issue #749
Provide the capability for the ResumableUpload class to resume the
upload that was in-progress when an upload program has halted and is
subsequently restarted.

Event UploadSessionData is called when an UploadUri is created to allow
the application to save the UploadUri to persistent storage.

Overloaded the Resume() and ResumeAsync() functions with the addition of
an UploadUri parameter to allow resuming an upload after a program
restart.

Resuming on program restart is only supported when using a seekable
ContentStream. Therefore, throw a NotImplementedException
if an resume upload is attempted after a program restart for a
non-seekable ContentStream.