Introduce a TokenRefreshManager to perform token refreshes
This is then used by ServiceCredential and UserCredential, which
effectively handle token refreshes the same way. In particular,
neither of them use the authUri parameter by default.
(ServiceAccountCredential does, by overriding
GetAccessTokenForRequest, but that's okay as it will only delegate
when the auth URI doesn't matter.)

The TokenRefreshManager avoids multiple concurrent requests by
remembering an in-flight task and using that for subsequent
requests.

Additionally, it reduces latency by starting a refresh task even
when the existing token is still valid, 6 minutes before expiry. We
deem a token that expires in 5 minutes to be too close to expiry to
use, so as long as there's a reasonable stream of requests and the
refresh is short, we should always be able to use a cached access
token.