Allow null Scopes in ExchangeCodeForTokenAsync() (#1488)
* Allow null Scopes

Previously, using an initializer with Scopes set to null would crash
when CreateAuthorizationCodeRequest() was called. Now, the null Scopes
propogates safely through as a null Scope for AuthorizationCodeRequestUrl.

This is necessary because some OAuth sites do not recognize the "scope"
parameter and will return an error instead of the authorization token if
the URL contains it.

I have updated the tests and added one for the default authorization
code request.

* Allow null Scopes in ExchangeCodeForTokenAsync()

Added missing code and test for the ExchangeCodeForTokenAsync() method
as well.