Allow post-generation tweaks on a per-API basis
The first API to need this is AdMob V1.

This fixes #1507 - or at least, will when the API is next published.
(That will be either when the minor version is bumped, or when the
API discovery doc changes. It would be relatively tricky to "force"
a change here.)

We don't want to do much per-API work, but tiny little tweaks like
this provide a lot of value for little work.