First pass of an extension method for requests
I'd want to write a few unit tests and then documentation.
The integration test shows how this can be used, including a
demonstration of "initialize request, set a property, execute the
request" all within a single expression.

It's unfortunate that this has to be an extension method, but that's
the only way to get the type inference working appropriately.

This is an alternative approach to the feature request in #1562.