Rewrite the Discovery Doc fetcher in C#
The only difference between the result of this and the result of
running the old get_discovery_documents.py script is that the script
version included some trailing spaces at the end of some lines.

The results aren't totally stable, as for some APIs where new
Discovery Docs are rolling out, the same request sometimes gives the
old revision and sometimes gives a new revision - that can be seen
in the browser though, so isn't really a code artifact.

With this change in place, this repository doesn't use Python at all.