feat: Makes extra parameters available in AuthorizationCodeResponseUrl
Some OAUTH providers have additional parameters in the Auth Code response that are necessary to function. (Specifically, Azure AD has parameters that dictate which azure cloud the code is to used on. I.E. Public, USGov, Germany, China, etc)  This change would allow people to not need to reproduce the functionality of LocalHostCodeReceiver in order to read these values.