Add target for .NET 4.6.1
Fixes #1871 (for modern platforms)

Fixes #1788

This *only* adds a target for Google.Apis.Auth, not Google.Apis or
Google.Apis.Core. We're relying on NuGet target resolution to
resolve to the correct versions for transitive dependencies - so
even if someone depends on Google.Apis (which doesn't have a .NET
4.6.1 target) from a .NET 4.6.1 application, they should still get
the .NET 4.6.1 version of Google.Apis.Auth.