chore: Configure Renovate for repository
This ignores the generated code, but does include both tools and
support. We may well find we *don't* want to accept a lot of the
changes suggested for the support libraries, but it's probably a
reasonable idea to get a first set of suggestions and consciously
choose whether or not to accept the change on a per-package basis.
If this becomes noisy in the longer term, we can exclude the support
libraries too (but ideally not the tests).