chore: Only write content when we have something to write
This appears to fix the issues we've seen in GitHub; the current theory is that the HttpClient may close the stream if it sees a content length of 0, before we get as far as writing the empty data. That write still checks whether the connection is open, causing a failure.