chore: Remove the Google.Apis.Auth.Mvc package
We won't delist this, but we won't publish new versions of it.

We don't have any tests for it; it's unclear whether it's actually useful; it target net45 which we'd like to remove support for.

(We'll remove the fairly-minimal documentation we have for it, but as the package is still present it does no harm anyway.)

If customers request that we reinstate it, we can do so.