feat: Makes batch response handling more efficient for big responses.
By disposing of the batch System.Net.HttpResponseMessage as soon as possible so that it, and more importantly, its contents, may be garbage collected if needed.

This is useful for big response of batch operations, since the whole response is read into memory and then split into individual responses that are then deserialized. Since we keep the whole response string in memory, we don't need to actual response and it's content buffer, which also contains the whole response, may be garbage collected if needed.

Closes #2603